simpai/requirement.txt
asgiref==3.7.2
asgiref==3.8.1
dj_database_url==2.3.0
dj_rest_auth==7.0.1
Django==4.2.6
Django==5.1.4
django_allauth==65.3.1
djangorestframework==3.15.2
gTTS==2.5.4
matplotlib==3.10.1
numpy==2.2.4
openai==1.68.2
openpyxl==3.1.5
pandas==2.2.3
plotly==5.24.1
PyPDF2==3.0.1
python-decouple==3.8
python_docx==1.1.2
reportlab==4.2.5
Requests==2.32.3
seaborn==0.13.2
stripe==11.5.0
